Network.AWS.EFS

Description

Amazon Elastic File System

Amazon Elastic File System (Amazon EFS) provides simple, scalable file storage for use with Amazon EC2 instances in the AWS Cloud. With Amazon EFS, storage capacity is elastic, growing and shrinking automatically as you add and remove files, so your applications have the storage they need, when they need it. For more information, see the User Guide .

Service Configuration

efs :: Service

API version 2015-02-01 of the Amazon Elastic File System SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by EFS.

MountTargetNotFound

_MountTargetNotFound :: AsError a => Getting (First ServiceError) a ServiceError

Returned if there is no mount target with the specified ID found in the caller's account.

SecurityGroupLimitExceeded

_SecurityGroupLimitExceeded :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the size of SecurityGroups specified in the request is greater than five.

SecurityGroupNotFound

_SecurityGroupNotFound :: AsError a => Getting (First ServiceError) a ServiceError

Returned if one of the specified security groups does not exist in the subnet's VPC.

MountTargetConflict

_MountTargetConflict :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the mount target would violate one of the specified restrictions based on the file system's existing mount targets.

UnsupportedAvailabilityZone

_UnsupportedAvailabilityZone :: AsError a => Getting (First ServiceError) a ServiceError

FileSystemLimitExceeded

_FileSystemLimitExceeded :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the AWS account has already created maximum number of file systems allowed per account.

NetworkInterfaceLimitExceeded

_NetworkInterfaceLimitExceeded :: AsError a => Getting (First ServiceError) a ServiceError

The calling account has reached the ENI limit for the specific AWS region. Client should try to delete some ENIs or get its account limit raised. For more information, see Amazon VPC Limits in the Amazon Virtual Private Cloud User Guide (see the Network interfaces per VPC entry in the table).

FileSystemAlreadyExists

_FileSystemAlreadyExists :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the file system you are trying to create already exists, with the creation token you provided.

SubnetNotFound

_SubnetNotFound :: AsError a => Getting (First ServiceError) a ServiceError

Returned if there is no subnet with ID SubnetId provided in the request.

FileSystemNotFound

_FileSystemNotFound :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the specified FileSystemId does not exist in the requester's AWS account.

IncorrectFileSystemLifeCycleState

_IncorrectFileSystemLifeCycleState :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the file system's life cycle state is not "created".

BadRequest

_BadRequest :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the request is malformed or contains an error such as an invalid parameter value or a missing required parameter.

NoFreeAddressesInSubnet

_NoFreeAddressesInSubnet :: AsError a => Getting (First ServiceError) a ServiceError

Returned if IpAddress was not specified in the request and there are no free IP addresses in the subnet.

DependencyTimeout

_DependencyTimeout :: AsError a => Getting (First ServiceError) a ServiceError

The service timed out trying to fulfill the request, and the client should try the call again.

FileSystemInUse

_FileSystemInUse :: AsError a => Getting (First ServiceError) a ServiceError

Returned if a file system has mount targets.

IncorrectMountTargetState

_IncorrectMountTargetState :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the mount target is not in the correct state for the operation.

InternalServerError

_InternalServerError :: AsError a => Getting (First ServiceError) a ServiceError

Returned if an error occurred on the server side.

IPAddressInUse

_IPAddressInUse :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the request specified an IpAddress that is already in use in the subnet.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

DescribeTags (Paginated)

module Network.AWS.EFS.DescribeTags

DescribeMountTargets (Paginated)

module Network.AWS.EFS.DescribeMountTargets

DescribeFileSystems (Paginated)

module Network.AWS.EFS.DescribeFileSystems

DeleteMountTarget

module Network.AWS.EFS.DeleteMountTarget

CreateTags

module Network.AWS.EFS.CreateTags

DeleteTags

module Network.AWS.EFS.DeleteTags

DescribeMountTargetSecurityGroups

module Network.AWS.EFS.DescribeMountTargetSecurityGroups

ModifyMountTargetSecurityGroups

module Network.AWS.EFS.ModifyMountTargetSecurityGroups

CreateFileSystem

module Network.AWS.EFS.CreateFileSystem

DeleteFileSystem

module Network.AWS.EFS.DeleteFileSystem

CreateMountTarget

module Network.AWS.EFS.CreateMountTarget

Types

LifeCycleState

data LifeCycleState

Constructors

Available

Creating

Deleted

Deleting

PerformanceMode

data PerformanceMode

Constructors

GeneralPurpose

MaxIO

FileSystemDescription

data FileSystemDescription

Description of the file system.

See: fileSystemDescription smart constructor.

fileSystemDescription

Arguments

:: Text
fsdOwnerId

-> Text
fsdCreationToken

-> Text
fsdFileSystemId

-> UTCTime
fsdCreationTime

-> LifeCycleState
fsdLifeCycleState

-> Natural
fsdNumberOfMountTargets

-> FileSystemSize
fsdSizeInBytes

-> PerformanceMode
fsdPerformanceMode

-> FileSystemDescription

Creates a value of FileSystemDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fsdEncrypted - A boolean value that, if true, indicates that the file system is encrypted.

fsdKMSKeyId - The id of an AWS Key Management Service (AWS KMS) customer master key (CMK) that was used to protect the encrypted file system.

fsdName - You can add tags to a file system, including a Name tag. For more information, see CreateTags . If the file system has a Name tag, Amazon EFS returns the value in this field.

fsdOwnerId - AWS account that created the file system. If the file system was created by an IAM user, the parent account to which the user belongs is the owner.

fsdCreationToken - Opaque string specified in the request.

fsdFileSystemId - ID of the file system, assigned by Amazon EFS.

fsdCreationTime - Time that the file system was created, in seconds (since 1970-01-01T00:00:00Z).

fsdLifeCycleState - Lifecycle phase of the file system.

fsdNumberOfMountTargets - Current number of mount targets that the file system has. For more information, see CreateMountTarget .

fsdSizeInBytes - Latest known metered size (in bytes) of data stored in the file system, in bytes, in its Value field, and the time at which that size was determined in its Timestamp field. The Timestamp value is the integer number of seconds since 1970-01-01T00:00:00Z. Note that the value does not represent the size of a consistent snapshot of the file system, but it is eventually consistent when there are no writes to the file system. That is, the value will represent actual size only if the file system is not modified for a period longer than a couple of hours. Otherwise, the value is not the exact size the file system was at any instant in time.

fsdPerformanceMode - The PerformanceMode of the file system.

fsdEncrypted :: Lens' FileSystemDescription (Maybe Bool)

A boolean value that, if true, indicates that the file system is encrypted.

fsdKMSKeyId :: Lens' FileSystemDescription (Maybe Text)

The id of an AWS Key Management Service (AWS KMS) customer master key (CMK) that was used to protect the encrypted file system.

fsdName :: Lens' FileSystemDescription (Maybe Text)

You can add tags to a file system, including a Name tag. For more information, see CreateTags . If the file system has a Name tag, Amazon EFS returns the value in this field.

fsdOwnerId :: Lens' FileSystemDescription Text

AWS account that created the file system. If the file system was created by an IAM user, the parent account to which the user belongs is the owner.

fsdCreationToken :: Lens' FileSystemDescription Text

Opaque string specified in the request.

fsdFileSystemId :: Lens' FileSystemDescription Text

ID of the file system, assigned by Amazon EFS.

fsdCreationTime :: Lens' FileSystemDescription UTCTime

Time that the file system was created, in seconds (since 1970-01-01T00:00:00Z).

fsdLifeCycleState :: Lens' FileSystemDescription LifeCycleState

Lifecycle phase of the file system.

fsdNumberOfMountTargets :: Lens' FileSystemDescription Natural

Current number of mount targets that the file system has. For more information, see CreateMountTarget .

fsdSizeInBytes :: Lens' FileSystemDescription FileSystemSize

Latest known metered size (in bytes) of data stored in the file system, in bytes, in its Value field, and the time at which that size was determined in its Timestamp field. The Timestamp value is the integer number of seconds since 1970-01-01T00:00:00Z. Note that the value does not represent the size of a consistent snapshot of the file system, but it is eventually consistent when there are no writes to the file system. That is, the value will represent actual size only if the file system is not modified for a period longer than a couple of hours. Otherwise, the value is not the exact size the file system was at any instant in time.

fsdPerformanceMode :: Lens' FileSystemDescription PerformanceMode

The PerformanceMode of the file system.

FileSystemSize

data FileSystemSize

Latest known metered size (in bytes) of data stored in the file system, in its Value field, and the time at which that size was determined in its Timestamp field. Note that the value does not represent the size of a consistent snapshot of the file system, but it is eventually consistent when there are no writes to the file system. That is, the value will represent the actual size only if the file system is not modified for a period longer than a couple of hours. Otherwise, the value is not necessarily the exact size the file system was at any instant in time.

See: fileSystemSize smart constructor.

fileSystemSize

Arguments

:: Natural
fssValue

-> FileSystemSize

Creates a value of FileSystemSize with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fssTimestamp - Time at which the size of data, returned in the Value field, was determined. The value is the integer number of seconds since 1970-01-01T00:00:00Z.

fssValue - Latest known metered size (in bytes) of data stored in the file system.

fssTimestamp :: Lens' FileSystemSize (Maybe UTCTime)

Time at which the size of data, returned in the Value field, was determined. The value is the integer number of seconds since 1970-01-01T00:00:00Z.

fssValue :: Lens' FileSystemSize Natural

Latest known metered size (in bytes) of data stored in the file system.

MountTargetDescription

data MountTargetDescription

Provides a description of a mount target.

See: mountTargetDescription smart constructor.

mountTargetDescription

Arguments

:: Text
mtdMountTargetId

-> Text
mtdFileSystemId

-> Text
mtdSubnetId

-> LifeCycleState
mtdLifeCycleState

-> MountTargetDescription

Creates a value of MountTargetDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mtdIPAddress - Address at which the file system may be mounted via the mount target.

mtdNetworkInterfaceId - ID of the network interface that Amazon EFS created when it created the mount target.

mtdOwnerId - AWS account ID that owns the resource.

mtdMountTargetId - System-assigned mount target ID.

mtdFileSystemId - ID of the file system for which the mount target is intended.

mtdSubnetId - ID of the mount target's subnet.

mtdLifeCycleState - Lifecycle state of the mount target.

mtdIPAddress :: Lens' MountTargetDescription (Maybe Text)

Address at which the file system may be mounted via the mount target.

mtdNetworkInterfaceId :: Lens' MountTargetDescription (Maybe Text)

ID of the network interface that Amazon EFS created when it created the mount target.

mtdOwnerId :: Lens' MountTargetDescription (Maybe Text)

AWS account ID that owns the resource.

mtdMountTargetId :: Lens' MountTargetDescription Text

System-assigned mount target ID.

mtdFileSystemId :: Lens' MountTargetDescription Text

ID of the file system for which the mount target is intended.

mtdSubnetId :: Lens' MountTargetDescription Text

ID of the mount target's subnet.

mtdLifeCycleState :: Lens' MountTargetDescription LifeCycleState

Lifecycle state of the mount target.

Tag

data Tag

A tag is a key-value pair. Allowed characters: letters, whitespace, and numbers, representable in UTF-8, and the following characters:+ - = . _ : /

See: tag smart constructor.

tag

Arguments

:: Text
tagKey

-> Text
tagValue

-> Tag

Creates a value of Tag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tagKey - Tag key (String). The key can't start with aws: .

tagValue - Value of the tag key.

tagKey :: Lens' Tag Text

Tag key (String). The key can't start with aws: .

tagValue :: Lens' Tag Text

Value of the tag key.