Network.AWS.EFS.CreateFileSystem

Description

Creates a new, empty file system. The operation requires a creation token in the request that Amazon EFS uses to ensure idempotent creation (calling the operation with same creation token has no effect). If a file system does not currently exist that is owned by the caller's AWS account with the specified creation token, this operation does the following:

Creates a new, empty file system. The file system will have an Amazon EFS assigned ID, and an initial lifecycle state creating .

Returns with the description of the created file system.

Otherwise, this operation returns a FileSystemAlreadyExists error with the ID of the existing file system.

The idempotent operation allows you to retry a CreateFileSystem call without risk of creating an extra file system. This can happen when an initial call fails in a way that leaves it uncertain whether or not a file system was actually created. An example might be that a transport level timeout occurred or your connection was reset. As long as you use the same creation token, if the initial call had succeeded in creating a file system, the client can learn of its existence from the FileSystemAlreadyExists error.

This operation also takes an optional PerformanceMode parameter that you choose for your file system. We recommend generalPurpose performance mode for most file systems. File systems using the maxIO performance mode can scale to higher levels of aggregate throughput and operations per second with a tradeoff of slightly higher latencies for most file operations. The performance mode can't be changed after the file system has been created. For more information, see Amazon EFS: Performance Modes .

After the file system is fully created, Amazon EFS sets its lifecycle state to available , at which point you can create one or more mount targets for the file system in your VPC. For more information, see CreateMountTarget . You mount your Amazon EFS file system on an EC2 instances in your VPC via the mount target. For more information, see Amazon EFS: How it Works .

This operation requires permissions for the elasticfilesystem:CreateFileSystem action.

Creating a Request

createFileSystem

Arguments

:: Text
cfsCreationToken

-> CreateFileSystem

Creates a value of CreateFileSystem with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cfsPerformanceMode - The PerformanceMode of the file system. We recommend generalPurpose performance mode for most file systems. File systems using the maxIO performance mode can scale to higher levels of aggregate throughput and operations per second with a tradeoff of slightly higher latencies for most file operations. This can't be changed after the file system has been created.

cfsEncrypted - A boolean value that, if true, creates an encrypted file system. When creating an encrypted file system, you have the option of specifying a 'CreateFileSystemRequest$KmsKeyId' for an existing AWS Key Management Service (AWS KMS) customer master key (CMK). If you don't specify a CMK, then the default CMK for Amazon EFS, awselasticfilesystem , is used to protect the encrypted file system.

cfsKMSKeyId - The id of the AWS KMS CMK that will be used to protect the encrypted file system. This parameter is only required if you want to use a non-default CMK. If this parameter is not specified, the default CMK for Amazon EFS is used. This id can be in one of the following formats: * Key ID - A unique identifier of the key. For example, 1234abcd-12ab-34cd-56ef-1234567890ab . * ARN - An Amazon Resource Name for the key. For example, arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab . * Key alias - A previously created display name for a key. For example, alias/projectKey1 . * Key alias ARN - An Amazon Resource Name for a key alias. For example, arn:aws:kms:us-west-2:444455556666:alias/projectKey1 . Note that if the KmsKeyId is specified, the 'CreateFileSystemRequest$Encrypted' parameter must be set to true.

cfsCreationToken - String of up to 64 ASCII characters. Amazon EFS uses this to ensure idempotent creation.

data CreateFileSystem

See: createFileSystem smart constructor.

Request Lenses

cfsPerformanceMode :: Lens' CreateFileSystem (Maybe PerformanceMode)

The PerformanceMode of the file system. We recommend generalPurpose performance mode for most file systems. File systems using the maxIO performance mode can scale to higher levels of aggregate throughput and operations per second with a tradeoff of slightly higher latencies for most file operations. This can't be changed after the file system has been created.

cfsEncrypted :: Lens' CreateFileSystem (Maybe Bool)

A boolean value that, if true, creates an encrypted file system. When creating an encrypted file system, you have the option of specifying a 'CreateFileSystemRequest$KmsKeyId' for an existing AWS Key Management Service (AWS KMS) customer master key (CMK). If you don't specify a CMK, then the default CMK for Amazon EFS, awselasticfilesystem , is used to protect the encrypted file system.

cfsKMSKeyId :: Lens' CreateFileSystem (Maybe Text)

The id of the AWS KMS CMK that will be used to protect the encrypted file system. This parameter is only required if you want to use a non-default CMK. If this parameter is not specified, the default CMK for Amazon EFS is used. This id can be in one of the following formats: * Key ID - A unique identifier of the key. For example, 1234abcd-12ab-34cd-56ef-1234567890ab . * ARN - An Amazon Resource Name for the key. For example, arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab . * Key alias - A previously created display name for a key. For example, alias/projectKey1 . * Key alias ARN - An Amazon Resource Name for a key alias. For example, arn:aws:kms:us-west-2:444455556666:alias/projectKey1 . Note that if the KmsKeyId is specified, the 'CreateFileSystemRequest$Encrypted' parameter must be set to true.

cfsCreationToken :: Lens' CreateFileSystem Text

String of up to 64 ASCII characters. Amazon EFS uses this to ensure idempotent creation.

Destructuring the Response

fileSystemDescription

Arguments

:: Text
fsdOwnerId

-> Text
fsdCreationToken

-> Text
fsdFileSystemId

-> UTCTime
fsdCreationTime

-> LifeCycleState
fsdLifeCycleState

-> Natural
fsdNumberOfMountTargets

-> FileSystemSize
fsdSizeInBytes

-> PerformanceMode
fsdPerformanceMode

-> FileSystemDescription

Creates a value of FileSystemDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fsdEncrypted - A boolean value that, if true, indicates that the file system is encrypted.

fsdKMSKeyId - The id of an AWS Key Management Service (AWS KMS) customer master key (CMK) that was used to protect the encrypted file system.

fsdName - You can add tags to a file system, including a Name tag. For more information, see CreateTags . If the file system has a Name tag, Amazon EFS returns the value in this field.

fsdOwnerId - AWS account that created the file system. If the file system was created by an IAM user, the parent account to which the user belongs is the owner.

fsdCreationToken - Opaque string specified in the request.

fsdFileSystemId - ID of the file system, assigned by Amazon EFS.

fsdCreationTime - Time that the file system was created, in seconds (since 1970-01-01T00:00:00Z).

fsdLifeCycleState - Lifecycle phase of the file system.

fsdNumberOfMountTargets - Current number of mount targets that the file system has. For more information, see CreateMountTarget .

fsdSizeInBytes - Latest known metered size (in bytes) of data stored in the file system, in bytes, in its Value field, and the time at which that size was determined in its Timestamp field. The Timestamp value is the integer number of seconds since 1970-01-01T00:00:00Z. Note that the value does not represent the size of a consistent snapshot of the file system, but it is eventually consistent when there are no writes to the file system. That is, the value will represent actual size only if the file system is not modified for a period longer than a couple of hours. Otherwise, the value is not the exact size the file system was at any instant in time.

fsdPerformanceMode - The PerformanceMode of the file system.

data FileSystemDescription

Description of the file system.

See: fileSystemDescription smart constructor.

Response Lenses

fsdEncrypted :: Lens' FileSystemDescription (Maybe Bool)

A boolean value that, if true, indicates that the file system is encrypted.

fsdKMSKeyId :: Lens' FileSystemDescription (Maybe Text)

The id of an AWS Key Management Service (AWS KMS) customer master key (CMK) that was used to protect the encrypted file system.

fsdName :: Lens' FileSystemDescription (Maybe Text)

You can add tags to a file system, including a Name tag. For more information, see CreateTags . If the file system has a Name tag, Amazon EFS returns the value in this field.

fsdOwnerId :: Lens' FileSystemDescription Text

AWS account that created the file system. If the file system was created by an IAM user, the parent account to which the user belongs is the owner.

fsdCreationToken :: Lens' FileSystemDescription Text

Opaque string specified in the request.

fsdFileSystemId :: Lens' FileSystemDescription Text

ID of the file system, assigned by Amazon EFS.

fsdCreationTime :: Lens' FileSystemDescription UTCTime

Time that the file system was created, in seconds (since 1970-01-01T00:00:00Z).

fsdLifeCycleState :: Lens' FileSystemDescription LifeCycleState

Lifecycle phase of the file system.

fsdNumberOfMountTargets :: Lens' FileSystemDescription Natural

Current number of mount targets that the file system has. For more information, see CreateMountTarget .

fsdSizeInBytes :: Lens' FileSystemDescription FileSystemSize

Latest known metered size (in bytes) of data stored in the file system, in bytes, in its Value field, and the time at which that size was determined in its Timestamp field. The Timestamp value is the integer number of seconds since 1970-01-01T00:00:00Z. Note that the value does not represent the size of a consistent snapshot of the file system, but it is eventually consistent when there are no writes to the file system. That is, the value will represent actual size only if the file system is not modified for a period longer than a couple of hours. Otherwise, the value is not the exact size the file system was at any instant in time.

fsdPerformanceMode :: Lens' FileSystemDescription PerformanceMode

The PerformanceMode of the file system.