Network.AWS.EFS.CreateMountTarget

Description

Creates a mount target for a file system. You can then mount the file system on EC2 instances via the mount target.

You can create one mount target in each Availability Zone in your VPC. All EC2 instances in a VPC within a given Availability Zone share a single mount target for a given file system. If you have multiple subnets in an Availability Zone, you create a mount target in one of the subnets. EC2 instances do not need to be in the same subnet as the mount target in order to access their file system. For more information, see Amazon EFS: How it Works .

In the request, you also specify a file system ID for which you are creating the mount target and the file system's lifecycle state must be available . For more information, see DescribeFileSystems .

In the request, you also provide a subnet ID, which determines the following:

VPC in which Amazon EFS creates the mount target

Availability Zone in which Amazon EFS creates the mount target

IP address range from which Amazon EFS selects the IP address of the mount target (if you don't specify an IP address in the request)

After creating the mount target, Amazon EFS returns a response that includes, a MountTargetId and an IpAddress . You use this IP address when mounting the file system in an EC2 instance. You can also use the mount target's DNS name when mounting the file system. The EC2 instance on which you mount the file system via the mount target can resolve the mount target's DNS name to its IP address. For more information, see How it Works: Implementation Overview .

Note that you can create mount targets for a file system in only one VPC, and there can be only one mount target per Availability Zone. That is, if the file system already has one or more mount targets created for it, the subnet specified in the request to add another mount target must meet the following requirements:

Must belong to the same VPC as the subnets of the existing mount targets

Must not be in the same Availability Zone as any of the subnets of the existing mount targets

If the request satisfies the requirements, Amazon EFS does the following:

Creates a new mount target in the specified subnet.

Also creates a new network interface in the subnet as follows:

If the request provides an IpAddress , Amazon EFS assigns that IP address to the network interface. Otherwise, Amazon EFS assigns a free address in the subnet (in the same way that the Amazon EC2 CreateNetworkInterface call does when a request does not specify a primary private IP address).

If the request provides SecurityGroups , this network interface is associated with those security groups. Otherwise, it belongs to the default security group for the subnet's VPC.

Assigns the description Mount target fsmt-id for file system fs-id where fsmt-id is the mount target ID, and fs-id is the FileSystemId .

Sets the requesterManaged property of the network interface to true , and the requesterId value to EFS .

Each Amazon EFS mount target has one corresponding requester-managed EC2 network interface. After the network interface is created, Amazon EFS sets the NetworkInterfaceId field in the mount target's description to the network interface ID, and the IpAddress field to its address. If network interface creation fails, the entire CreateMountTarget operation fails.

We recommend you create a mount target in each of the Availability Zones. There are cost considerations for using a file system in an Availability Zone through a mount target created in another Availability Zone. For more information, see Amazon EFS . In addition, by always using a mount target local to the instance's Availability Zone, you eliminate a partial failure scenario. If the Availability Zone in which your mount target is created goes down, then you won't be able to access your file system through that mount target.

This operation requires permissions for the following action on the file system:

elasticfilesystem:CreateMountTarget

This operation also requires permissions for the following Amazon EC2 actions:

ec2:DescribeSubnets

ec2:DescribeNetworkInterfaces

ec2:CreateNetworkInterface

Creating a Request

createMountTarget

Arguments

:: Text
cmtFileSystemId

-> Text
cmtSubnetId

-> CreateMountTarget

Creates a value of CreateMountTarget with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cmtIPAddress - Valid IPv4 address within the address range of the specified subnet.

cmtSecurityGroups - Up to five VPC security group IDs, of the form sg-xxxxxxxx . These must be for the same VPC as subnet specified.

cmtFileSystemId - ID of the file system for which to create the mount target.

cmtSubnetId - ID of the subnet to add the mount target in.

data CreateMountTarget

See: createMountTarget smart constructor.

Request Lenses

cmtIPAddress :: Lens' CreateMountTarget (Maybe Text)

Valid IPv4 address within the address range of the specified subnet.

cmtSecurityGroups :: Lens' CreateMountTarget [Text]

Up to five VPC security group IDs, of the form sg-xxxxxxxx . These must be for the same VPC as subnet specified.

cmtFileSystemId :: Lens' CreateMountTarget Text

ID of the file system for which to create the mount target.

cmtSubnetId :: Lens' CreateMountTarget Text

ID of the subnet to add the mount target in.

Destructuring the Response

mountTargetDescription

Arguments

:: Text
mtdMountTargetId

-> Text
mtdFileSystemId

-> Text
mtdSubnetId

-> LifeCycleState
mtdLifeCycleState

-> MountTargetDescription

Creates a value of MountTargetDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mtdIPAddress - Address at which the file system may be mounted via the mount target.

mtdNetworkInterfaceId - ID of the network interface that Amazon EFS created when it created the mount target.

mtdOwnerId - AWS account ID that owns the resource.

mtdMountTargetId - System-assigned mount target ID.

mtdFileSystemId - ID of the file system for which the mount target is intended.

mtdSubnetId - ID of the mount target's subnet.

mtdLifeCycleState - Lifecycle state of the mount target.

data MountTargetDescription

Provides a description of a mount target.

See: mountTargetDescription smart constructor.

Response Lenses

mtdIPAddress :: Lens' MountTargetDescription (Maybe Text)

Address at which the file system may be mounted via the mount target.

mtdNetworkInterfaceId :: Lens' MountTargetDescription (Maybe Text)

ID of the network interface that Amazon EFS created when it created the mount target.

mtdOwnerId :: Lens' MountTargetDescription (Maybe Text)

AWS account ID that owns the resource.

mtdMountTargetId :: Lens' MountTargetDescription Text

System-assigned mount target ID.

mtdFileSystemId :: Lens' MountTargetDescription Text

ID of the file system for which the mount target is intended.

mtdSubnetId :: Lens' MountTargetDescription Text

ID of the mount target's subnet.

mtdLifeCycleState :: Lens' MountTargetDescription LifeCycleState

Lifecycle state of the mount target.