Network.AWS.EFS.DeleteMountTarget

Description

Deletes the specified mount target.

This operation forcibly breaks any mounts of the file system via the mount target that is being deleted, which might disrupt instances or applications using those mounts. To avoid applications getting cut off abruptly, you might consider unmounting any mounts of the mount target, if feasible. The operation also deletes the associated network interface. Uncommitted writes may be lost, but breaking a mount target using this operation does not corrupt the file system itself. The file system you created remains. You can mount an EC2 instance in your VPC via another mount target.

This operation requires permissions for the following action on the file system:

elasticfilesystem:DeleteMountTarget

The operation also requires permissions for the following Amazon EC2 action on the mount target's network interface:

ec2:DeleteNetworkInterface

Creating a Request

deleteMountTarget

Arguments

:: Text
dMountTargetId

-> DeleteMountTarget

Creates a value of DeleteMountTarget with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dMountTargetId - ID of the mount target to delete (String).

data DeleteMountTarget

See: deleteMountTarget smart constructor.

Request Lenses

dMountTargetId :: Lens' DeleteMountTarget Text

ID of the mount target to delete (String).

Destructuring the Response

deleteMountTargetResponse :: DeleteMountTargetResponse

Creates a value of DeleteMountTargetResponse with the minimum fields required to make a request.

data DeleteMountTargetResponse

See: deleteMountTargetResponse smart constructor.