Network.AWS.EFS.DescribeFileSystems

Description

Returns the description of a specific Amazon EFS file system if either the file system CreationToken or the FileSystemId is provided. Otherwise, it returns descriptions of all file systems owned by the caller's AWS account in the AWS Region of the endpoint that you're calling.

When retrieving all file system descriptions, you can optionally specify the MaxItems parameter to limit the number of descriptions in a response. If more file system descriptions remain, Amazon EFS returns a NextMarker , an opaque token, in the response. In this case, you should send a subsequent request with the Marker request parameter set to the value of NextMarker .

To retrieve a list of your file system descriptions, this operation is used in an iterative process, where DescribeFileSystems is called first without the Marker and then the operation continues to call it with the Marker parameter set to the value of the NextMarker from the previous response until the response has no NextMarker .

The implementation may return fewer than MaxItems file system descriptions while still including a NextMarker value.

The order of file systems returned in the response of one DescribeFileSystems call and the order of file systems returned across the responses of a multi-call iteration is unspecified.

This operation requires permissions for the elasticfilesystem:DescribeFileSystems action.

This operation returns paginated results.

Creating a Request

describeFileSystems :: DescribeFileSystems

Creates a value of DescribeFileSystems with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dfsFileSystemId - (Optional) ID of the file system whose description you want to retrieve (String).

dfsCreationToken - (Optional) Restricts the list to the file system with this creation token (String). You specify a creation token when you create an Amazon EFS file system.

dfsMarker - (Optional) Opaque pagination token returned from a previous DescribeFileSystems operation (String). If present, specifies to continue the list from where the returning call had left off.

dfsMaxItems - (Optional) Specifies the maximum number of file systems to return in the response (integer). This parameter value must be greater than 0. The number of items that Amazon EFS returns is the minimum of the MaxItems parameter specified in the request and the service's internal maximum number of items per page.

data DescribeFileSystems

See: describeFileSystems smart constructor.

Request Lenses

dfsFileSystemId :: Lens' DescribeFileSystems (Maybe Text)

(Optional) ID of the file system whose description you want to retrieve (String).

dfsCreationToken :: Lens' DescribeFileSystems (Maybe Text)

(Optional) Restricts the list to the file system with this creation token (String). You specify a creation token when you create an Amazon EFS file system.

dfsMarker :: Lens' DescribeFileSystems (Maybe Text)

(Optional) Opaque pagination token returned from a previous DescribeFileSystems operation (String). If present, specifies to continue the list from where the returning call had left off.

dfsMaxItems :: Lens' DescribeFileSystems (Maybe Natural)

(Optional) Specifies the maximum number of file systems to return in the response (integer). This parameter value must be greater than 0. The number of items that Amazon EFS returns is the minimum of the MaxItems parameter specified in the request and the service's internal maximum number of items per page.

Destructuring the Response

describeFileSystemsResponse

Arguments

:: Int
dfsrsResponseStatus

-> DescribeFileSystemsResponse

Creates a value of DescribeFileSystemsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dfsrsFileSystems - Array of file system descriptions.

dfsrsMarker - Present if provided by caller in the request (String).

dfsrsNextMarker - Present if there are more file systems than returned in the response (String). You can use the NextMarker in the subsequent request to fetch the descriptions.

dfsrsResponseStatus - -- | The response status code.

data DescribeFileSystemsResponse

See: describeFileSystemsResponse smart constructor.

Response Lenses

dfsrsFileSystems :: Lens' DescribeFileSystemsResponse [FileSystemDescription]

Array of file system descriptions.

dfsrsMarker :: Lens' DescribeFileSystemsResponse (Maybe Text)

Present if provided by caller in the request (String).

dfsrsNextMarker :: Lens' DescribeFileSystemsResponse (Maybe Text)

Present if there are more file systems than returned in the response (String). You can use the NextMarker in the subsequent request to fetch the descriptions.

dfsrsResponseStatus :: Lens' DescribeFileSystemsResponse Int

- | The response status code.