Network.AWS.EFS.DescribeMountTargets

Description

Returns the descriptions of all the current mount targets, or a specific mount target, for a file system. When requesting all of the current mount targets, the order of mount targets returned in the response is unspecified.

This operation requires permissions for the elasticfilesystem:DescribeMountTargets action, on either the file system ID that you specify in FileSystemId , or on the file system of the mount target that you specify in MountTargetId .

This operation returns paginated results.

Creating a Request

describeMountTargets :: DescribeMountTargets

Creates a value of DescribeMountTargets with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmtFileSystemId - (Optional) ID of the file system whose mount targets you want to list (String). It must be included in your request if MountTargetId is not included.

dmtMarker - (Optional) Opaque pagination token returned from a previous DescribeMountTargets operation (String). If present, it specifies to continue the list from where the previous returning call left off.

dmtMaxItems - (Optional) Maximum number of mount targets to return in the response. It must be an integer with a value greater than zero.

dmtMountTargetId - (Optional) ID of the mount target that you want to have described (String). It must be included in your request if FileSystemId is not included.

data DescribeMountTargets

See: describeMountTargets smart constructor.

Request Lenses

dmtFileSystemId :: Lens' DescribeMountTargets (Maybe Text)

(Optional) ID of the file system whose mount targets you want to list (String). It must be included in your request if MountTargetId is not included.

dmtMarker :: Lens' DescribeMountTargets (Maybe Text)

(Optional) Opaque pagination token returned from a previous DescribeMountTargets operation (String). If present, it specifies to continue the list from where the previous returning call left off.

dmtMaxItems :: Lens' DescribeMountTargets (Maybe Natural)

(Optional) Maximum number of mount targets to return in the response. It must be an integer with a value greater than zero.

dmtMountTargetId :: Lens' DescribeMountTargets (Maybe Text)

(Optional) ID of the mount target that you want to have described (String). It must be included in your request if FileSystemId is not included.

Destructuring the Response

describeMountTargetsResponse

Arguments

:: Int
dmtrsResponseStatus

-> DescribeMountTargetsResponse

Creates a value of DescribeMountTargetsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmtrsMountTargets - Returns the file system's mount targets as an array of MountTargetDescription objects.

dmtrsMarker - If the request included the Marker , the response returns that value in this field.

dmtrsNextMarker - If a value is present, there are more mount targets to return. In a subsequent request, you can provide Marker in your request with this value to retrieve the next set of mount targets.

dmtrsResponseStatus - -- | The response status code.

data DescribeMountTargetsResponse

See: describeMountTargetsResponse smart constructor.

Response Lenses

dmtrsMountTargets :: Lens' DescribeMountTargetsResponse [MountTargetDescription]

Returns the file system's mount targets as an array of MountTargetDescription objects.

dmtrsMarker :: Lens' DescribeMountTargetsResponse (Maybe Text)

If the request included the Marker , the response returns that value in this field.

dmtrsNextMarker :: Lens' DescribeMountTargetsResponse (Maybe Text)

If a value is present, there are more mount targets to return. In a subsequent request, you can provide Marker in your request with this value to retrieve the next set of mount targets.

dmtrsResponseStatus :: Lens' DescribeMountTargetsResponse Int

- | The response status code.