Network.AWS.EFS.ModifyMountTargetSecurityGroups

Description

Modifies the set of security groups in effect for a mount target.

When you create a mount target, Amazon EFS also creates a new network interface. For more information, see CreateMountTarget . This operation replaces the security groups in effect for the network interface associated with a mount target, with the SecurityGroups provided in the request. This operation requires that the network interface of the mount target has been created and the lifecycle state of the mount target is not deleted .

The operation requires permissions for the following actions:

elasticfilesystem:ModifyMountTargetSecurityGroups action on the mount target's file system.

ec2:ModifyNetworkInterfaceAttribute action on the mount target's network interface.

Creating a Request

modifyMountTargetSecurityGroups

Arguments

:: Text
mmtsgMountTargetId

-> ModifyMountTargetSecurityGroups

Creates a value of ModifyMountTargetSecurityGroups with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mmtsgSecurityGroups - Array of up to five VPC security group IDs.

mmtsgMountTargetId - ID of the mount target whose security groups you want to modify.

data ModifyMountTargetSecurityGroups

See: modifyMountTargetSecurityGroups smart constructor.

Request Lenses

mmtsgSecurityGroups :: Lens' ModifyMountTargetSecurityGroups [Text]

Array of up to five VPC security group IDs.

mmtsgMountTargetId :: Lens' ModifyMountTargetSecurityGroups Text

ID of the mount target whose security groups you want to modify.

Destructuring the Response

modifyMountTargetSecurityGroupsResponse :: ModifyMountTargetSecurityGroupsResponse

Creates a value of ModifyMountTargetSecurityGroupsResponse with the minimum fields required to make a request.

data ModifyMountTargetSecurityGroupsResponse

See: modifyMountTargetSecurityGroupsResponse smart constructor.